+++ Runtime/Scripts/PrimitiveCreateContextBase.cs
@@ -6,7 +6,7 @@ namespace GLTFast {
     
     abstract class PrimitiveCreateContextBase {
         public int primtiveIndex;
-        public MeshPrimitive primitive;
+        public int[] materials;
         public abstract bool IsCompleted {get;}
         public abstract Primitive? CreatePrimitive();
     }