+++ Runtime/Scripts/VertexAttributeConfig.cs
@@ -1,0 +1,16 @@
+﻿#if UNITY_EDITOR
+
+using System.Collections.Generic;
+
+namespace GLTFast
+{
+    class VertexAttributeConfig
+    {
+        public HashSet<int> meshIndices;
+
+        public VertexAttributeConfig() {
+            meshIndices = new HashSet<int>();
+        }
+    }
+}
+#endif